Implement bounded key type and O(1) callback list access
Improve type-safety of the event type by wrapping an integer and
exposing a constant maximum ID. This enabled the use of an array for the
list of callbacks and the associated optimizations.

Also, fix the thread-safety issues introduced by the `AsyncExecutor`
before. Passing a reference to a `CallbackStore` instance is safe, as
the executor is always a member of `EventSystem`, so the reference will
never outlive the actual storage.

Signed-off-by: Florian Limberger <florian.limberger@inovex.de>